Recognizes translate intent, default language to translate from set to English, moved translation methods in agent class
